+++ crates/router/src/services/authorization/permission_groups.rs
@@ -144,8 +144,7 @@ impl ParentGroupExt for ParentGroup {
     ) -> Option<HashMap<Self, String>> {
         let descriptions_map = Self::iter()
             .filter_map(|parent| {
-                let has_groups = groups.iter().any(|group| group.parent() == parent);
-                if !has_groups {
+                if !groups.iter().any(|group| group.parent() == parent) {
                     return None;
                 }
                 let filtered_resources: Vec<_> = parent
